pyproject.toml
[project]
name = "begunici"
version = "0.1.0"
description = "Tim"
requires-python = ">=3.11"

dependencies = [
    # --- Web Frameworks ---
    "Django==4.2.15",
    "djangorestframework==3.15.2",
    "django-filter==24.3",
    "drf-spectacular==0.23.1",

    # --- Databases / ORM ---
    "psycopg2>=2.9",      # PostgreSQL driver
    "PyMySQL>=1.1",       # MySQL driver
    "py-postgresql>=1.2", # optional PostgreSQL client
    "sqlparse>=0.4",

    # --- Utilities ---
    "python-dotenv>=0.21",
    "stripe==4.1.0",
    "requests>=2.26",
    "python-dateutil>=2.8",
    "pytz",      # legacy, Django still uses it
    "tzdata",    # timezone info for Windows / py 3.9+
    "openpyxl>=3.1.5",   # Excel export

    # --- Optional ---
    "Flask==2.0.3",   # you had it in venv, unsure if actually needed
    "livereload>=2.6" # for hot reload dev server?
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "flake8>=3.8",
    "pylint>=2.6",
    "isort>=5.6",
    "ruff>=0.12.10"
]

[build-system]
requires = ["setuptools>=65", "wheel>=0.41"]
build-backend = "setuptools.build_meta"
